refactor: update common components imports to use shadcn directory
